refactor: 스터디노트 답글 위치변경
