+++ src/features/study-note-comment/components/comment-teacher-answer.tsx
@@ -1,3 +1,5 @@
+'use client';
+
 import { CommentAnswerCard } from './comment-answer-card';
 
 export const CommentTeacherAnswer = () => {
